Customer repository + Category id type fix

+++ src/main/java/com/example/wdmsystem/category/system/Category.java
@@ -14,11 +14,11 @@ public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public Long id;
+    public Integer id;
     public int merchantId;
     public String title;
 
-    public Category(Long id, int merchantId, String title) {
+    public Category(Integer id, int merchantId, String title) {
         this.id = id;
         this.merchantId = merchantId;
         this.title = title;
